move tests into public/tests and create helpers for opportunity and events
